requirements.txt
# Core dependencies
# Requires Python 3.9+

# Geospatial and network analysis
osmnx>=1.6.0
networkx>=3.0
shapely>=2.0.0
geopandas>=0.13.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.6
geoalchemy2>=0.14.0

# Optimization solvers
ortools>=9.6.0  # Free, open-source CP solver
# gurobipy>=10.0.0  # Optional: MILP solver (requires license)

# Visualization
folium>=0.14.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0
unittest-xml-reporting>=3.2.0

# Performance profiling
psutil>=5.9.5
memory-profiler>=0.61.0

# Development
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Documentation
sphinx>=6.2.0
sphinx-rtd-theme>=1.2.0

# Jupyter (optional, for analysis)
jupyter>=1.0.0
ipykernel>=6.23.0
